Changed leds to use charlieplexing
Leds use now charlieplexing which frees two pins for other use. That way
we can later add support for another button and a piezo speaker.
Technically it's possible to add one more led there if wanted.

Removed half minute display because it would be quite a pain to make
now. It might be added back later.

This implementation sets the timer to about 250 Hz using a trick. The
trick is to change the TOP value slightly so cycles are a always bit off
but their average is correct.

Leds are set by tables that define which pins are outputs and which
pins are high for each led.

After the timer runs out the MCU is kept in sleep mode as much as
possible (which is not a lot more really).